remove namespace,service from update
